connect: extract common vars to lib
cmd/connect is a subpackage with `tt connect` implementation, yet its
vars are reused in several other commands (and ETCD_* vars are never
used in `tt connect` at all). To improve reusing for [1], this patch
extracts them to lib/connect, which is assumed to be not just a
"tt connect library", but a "library related to connecting".

1. https://github.com/tarantool/tt-ee/issues/171
