+++ cli/connect/const.go
@@ -1,19 +1,5 @@
 package connect
 
-// TarantoolUsernameEnv is an environment variable with a username for
-// Tarantool.
-const TarantoolUsernameEnv = "TT_CLI_USERNAME"
-
-// TarantoolPasswordEnv is an environment variable with a password for
-// Tarantool.
-const TarantoolPasswordEnv = "TT_CLI_PASSWORD"
-
-// EtcdUsernameEnv is an environment variable with a username for etcd.
-const EtcdUsernameEnv = "TT_CLI_ETCD_USERNAME"
-
-// EtcdPasswordEnv is an environment variable with a password for etcd.
-const EtcdPasswordEnv = "TT_CLI_ETCD_PASSWORD"
-
 // setLanguagePrefix is used to set a language in the Tarantool console.
 const setLanguagePrefix = "\\set language"
